some small changes in preparation for more refactoring
