+++ waltid-libraries/protocols/waltid-openid4vc/src/commonMain/kotlin/id/walt/oid4vc/OpenID4VCI.kt
@@ -366,4 +366,231 @@ object OpenID4VCI {
     suspend fun generateDeferredCredentialToken(sessionId: String, issuer: String, credentialId: String, tokenKey: Key): String {
         return OpenID4VC.generateToken(sessionId, issuer, TokenTarget.DEFERRED_CREDENTIAL, credentialId, tokenKey)
     }
+//
+//    suspend private fun createCredentialResponseFor(sessionId: String, credentialResult: CredentialResult, openIDProviderMetadata: OpenIDProviderMetadata, sessionExpirationTime: Instant, tokenKey: Key): CredentialResponse {
+//        return credentialResult.credential?.let { credential ->
+//            CredentialResponse.success(credentialResult.format, credential, customParameters = credentialResult.customParameters)
+//        } ?: CredentialResponse.deferred(
+//                credentialResult.format,
+//                generateDeferredCredentialToken(sessionId,
+//                    openIDProviderMetadata.issuer ?: throw Exception("No issuer defined in provider metadata"),
+//                    credentialResult.credentialId ?: throw Exception("credentialId must not be null, if credential issuance is deferred."),
+//                    tokenKey),
+//                randomUUID(),
+//                sessionExpirationTime - Clock.System.now()
+//            )
+//        }
+//
+//    suspend fun generateCredentialResponse(sessionId: String, credentialRequest: CredentialRequest, openIDProviderMetadata: OpenIDProviderMetadata, sessionExpirationTime: Instant, nonce: String, tokenKey: Key): CredentialResponse {
+//        // access_token should be validated on API level and issuance session extracted
+//        // Validate credential request (proof of possession, etc)
+//        val validationResult = validateCredentialRequest(credentialRequest, nonce, openIDProviderMetadata)
+//        if(!validationResult.success)
+//            throw CredentialError(credentialRequest, CredentialErrorCode.invalid_request, message = validationResult.message)
+//        // create credential result
+//        val credentialResult = generateCredential(credentialRequest)
+//        return createCredentialResponseFor(sessionId, credentialResult, openIDProviderMetadata, sessionExpirationTime, tokenKey)
+//    }
+//
+//    suspend fun getValidatedDeferredCredentialId(acceptanceToken: String, openIDProviderMetadata: OpenIDProviderMetadata, tokenKey: Key): String {
+//        val accessInfo =
+//            OpenID4VC.verifyAndParseToken(acceptanceToken, openIDProviderMetadata.issuer!!, TokenTarget.DEFERRED_CREDENTIAL, tokenKey) ?: throw DeferredCredentialError(
+//                CredentialErrorCode.invalid_token,
+//                message = "Invalid acceptance token"
+//            )
+//        val credentialId = accessInfo[JWTClaims.Payload.jwtID]!!.jsonPrimitive.content
+//        return credentialId
+//    }
+//
+//    suspend fun generateCredential(credentialRequest: CredentialRequest): CredentialResult {
+//        log.debug { "GENERATING CREDENTIAL:" }
+//        log.debug { "Credential request: $credentialRequest" }
+//        log.debug { "CREDENTIAL REQUEST JSON -------:" }
+//        log.debug { Json.encodeToString(credentialRequest) }
+//
+//        return when (credentialRequest.format) {
+//            CredentialFormat.mso_mdoc -> doGenerateMDoc(credentialRequest)
+//            else -> doGenerateCredential(credentialRequest)
+//        }
+//    }
+//
+//    @OptIn(ExperimentalSerializationApi::class, ExperimentalStdlibApi::class)
+//    private suspend fun doGenerateCredential(
+//        credentialRequest: CredentialRequest,
+//        issuerKey: Key, issuerDid: String?,
+//        credentialData: JsonObject
+//    ): CredentialResult {
+//        if (credentialRequest.format == CredentialFormat.mso_mdoc) throw CredentialError(
+//            credentialRequest, CredentialErrorCode.unsupported_credential_format
+//        )
+//
+//        val proofHeader = credentialRequest.proof?.jwt?.let { JwtUtils.parseJWTHeader(it) } ?: throw CredentialError(
+//            credentialRequest, CredentialErrorCode.invalid_or_missing_proof, message = "Proof must be JWT proof"
+//        )
+//
+//        val holderKid = proofHeader[JWTClaims.Header.keyID]?.jsonPrimitive?.content
+//        val holderKey = proofHeader[JWTClaims.Header.jwk]?.jsonObject
+//
+//        if (holderKey.isNullOrEmpty() && holderKid.isNullOrEmpty()) throw CredentialError(
+//            credentialRequest,
+//            CredentialErrorCode.invalid_or_missing_proof,
+//            message = "Proof JWT header must contain kid or jwk claim"
+//        )
+//        val holderDid = if (!holderKid.isNullOrEmpty() && DidUtils.isDidUrl(holderKid)) holderKid.substringBefore("#") else null
+//        val nonce = OpenID4VCI.getNonceFromProof(credentialRequest.proof!!) ?: throw CredentialError(
+//            credentialRequest,
+//            CredentialErrorCode.invalid_or_missing_proof, message = "Proof must contain nonce"
+//        )
+//
+//        return CredentialResult(format = credentialRequest.format, credential = JsonPrimitive({
+//            val vc = data.request.credentialData ?: throw MissingFieldException(listOf("credentialData"), "credentialData")
+//
+//            var issuerKid = issuerDid ?: issuerKey.getKeyId()
+//            if(!issuerDid.isNullOrEmpty()) {
+//                if (issuerDid.startsWith("did:key") && issuerDid.length == 186) // EBSI conformance corner case when issuer uses did:key instead of did:ebsi and no trust framework is defined
+//                    issuerKid = issuerDid + "#" + issuerDid.removePrefix("did:key:")
+//                else if (issuerDid.startsWith("did:ebsi"))
+//                    issuerKid = issuerDid + "#" + issuerKey.getKeyId()
+//            }
+//
+//            val holderKeyJWK =  JWKKey.importJWK(holderKey.toString()).getOrNull()?.exportJWKObject()?.plus("kid" to JWKKey.importJWK(holderKey.toString()).getOrThrow().getKeyId())?.toJsonObject()
+//
+//            when (credentialRequest.format) {
+//                CredentialFormat.sd_jwt_vc -> sdJwtVc(
+//                    holderKeyJWK,
+//                    vc,
+//                    holderDid, issuerKid)
+//                else -> w3cSdJwtVc(W3CVC(vc), issuerKid, holderDid, holderKey)
+//            }
+//            log.debug { "Respond VC: $it" }
+//        }))
+//    }
+//
+//    private suspend fun w3cSdJwtVc(
+//        vc: W3CVC,
+//        issuerKid: String, issuerKey: Key, issuerDid: String?,
+//        holderDid: String?,
+//        mapping: JsonObject?, selectiveDisclosure: SDMap?
+//    ) = when(selectiveDisclosure.isNullOrEmpty()) {
+//        true -> vc.mergingJwtIssue(
+//            issuerKey = issuerKey,
+//            issuerDid = issuerDid,
+//            issuerKid = issuerKid,
+//            subjectDid = holderDid ?: "",
+//            mappings = mapping ?: JsonObject(emptyMap()),
+//            additionalJwtHeader = emptyMap(),
+//            additionalJwtOptions = emptyMap()
+//        )
+//        else -> vc.mergingSdJwtIssue(
+//            issuerKey = issuerKey,
+//            issuerDid = issuerDid,
+//            subjectDid = holderDid ?: "",
+//            mappings = mapping ?: JsonObject(emptyMap()),
+//            additionalJwtHeaders = emptyMap(),
+//            additionalJwtOptions = emptyMap(),
+//            disclosureMap = selectiveDisclosure
+//        )
+//    }.also {
+////        sendCallback("jwt_issue", buildJsonObject {
+////            put("jwt", it)
+////        })
+//    }
+//
+//    private suspend fun sdJwtVc(
+//        vc: JsonObject,
+//        issuerKid: String, issuerKey: Key, issuerDid: String?,
+//        holderDid: String?,
+//        mapping: JsonObject?, selectiveDisclosure: SDMap?,
+//        holderKey: JsonObject?
+//    ): String = id.walt.sdjwt.SDJwtVC.sign(
+//        sdPayload = id.walt.sdjwt.SDPayload.createSDPayload(
+//            vc.mergeSDJwtVCPayloadWithMapping(
+//                mapping = mapping ?: JsonObject(emptyMap()),
+//                context = mapOf(
+//                    "issuerDid" to issuerDid,
+//                    "subjectDid" to holderDid
+//                ).filterValues { !it.isNullOrEmpty() }.mapValues { JsonPrimitive(it.value) },
+//                id.walt.credentials.issuance.dataFunctions
+//            ),
+//            selectiveDisclosure ?: SDMap(mapOf())
+//        ),
+//        jwtCryptoProvider = SingleKe,
+//        issuerDid = (issuerDid ?: "").ifEmpty { issuerKey.getKeyId() },
+//        holderDid = holderDid,
+//        holderKeyJWK = holderKey,
+//        issuerKeyId = issuerKid,
+//        vct = metadata.credentialConfigurationsSupported?.get(request.credentialConfigurationId)?.vct ?: throw ConfigurationException(
+//            ConfigException("No vct configured for given credential configuration id: ${request.credentialConfigurationId}")
+//        ),
+//        additionalJwtHeader = request.x5Chain?.let {
+//            mapOf("x5c" to JsonArray(it.map { cert -> cert.toJsonElement() }))
+//        } ?: mapOf()
+//    ).toString().also {
+//        sendCallback("sdjwt_issue", buildJsonObject {
+//            put("sdjwt", it)
+//        })
+//    }
+//
+//    @OptIn(ExperimentalSerializationApi::class)
+//    private suspend fun doGenerateMDoc(
+//        credentialRequest: CredentialRequest,
+//    ): CredentialResult {
+//        val coseSign1 = Cbor.decodeFromByteArray<COSESign1>(
+//            credentialRequest.proof?.cwt?.base64UrlDecode() ?: throw CredentialError(
+//                credentialRequest,
+//                CredentialErrorCode.invalid_or_missing_proof, message = "No CWT proof found on credential request"
+//            )
+//        )
+//        val holderKey = COSESign1Utils.extractHolderKey(coseSign1)
+//        val nonce = OpenID4VCI.getNonceFromProof(credentialRequest.proof!!) ?: throw CredentialError(
+//            credentialRequest,
+//            CredentialErrorCode.invalid_or_missing_proof,
+//            message = "No nonce found on proof"
+//        )
+//        println("RETRIEVING VC FROM TOKEN MAPPING: $nonce")
+//        val data: IssuanceSessionData = tokenCredentialMapping[nonce]?.first()
+//            ?: throw CredentialError(
+//                credentialRequest,
+//                CredentialErrorCode.invalid_request,
+//                "The issuanceIdCredentialMapping does not contain a mapping for: $nonce!"
+//            )
+//        val issuerSignedItems = data.request.mdocData ?: throw MissingFieldException(listOf("mdocData"), "mdocData")
+//        val issuerKey = JWK.parse(runBlocking { data.issuerKey.key.exportJWK() }).toECKey()
+//        val keyId = runBlocking { data.issuerKey.key.getKeyId() }
+//        val cryptoProvider = SimpleCOSECryptoProvider(listOf(
+//            COSECryptoProviderKeyInfo(
+//                keyId, AlgorithmID.ECDSA_256, issuerKey.toECPublicKey(), issuerKey.toECPrivateKey(),
+//                x5Chain = data.request.x5Chain?.map { X509CertUtils.parse(it) } ?: listOf(),
+//                trustedRootCAs = data.request.trustedRootCAs?.map { X509CertUtils.parse(it) } ?: listOf()
+//            )
+//        ))
+//        val mdoc = MDocBuilder(
+//            credentialRequest.docType
+//                ?: throw CredentialError(
+//                    credentialRequest,
+//                    CredentialErrorCode.invalid_request,
+//                    message = "Missing doc type in credential request"
+//                )
+//        ).apply {
+//            issuerSignedItems.forEach { namespace ->
+//                namespace.value.forEach { property ->
+//                    addItemToSign(namespace.key, property.key, property.value.toDataElement())
+//                }
+//            }
+//        }.sign( // TODO: expiration date!
+//            ValidityInfo(Clock.System.now(), Clock.System.now(), Clock.System.now().plus(365 * 24, DateTimeUnit.HOUR)),
+//            DeviceKeyInfo(
+//                DataElement.fromCBOR(
+//                    OneKey(holderKey.publicKey, null).AsCBOR().EncodeToBytes()
+//                )
+//            ), cryptoProvider, keyId
+//        ).also {
+//            data.sendCallback("generated_mdoc", buildJsonObject { put("mdoc", it.toCBORHex()) })
+//        }
+//        return CredentialResult(
+//            CredentialFormat.mso_mdoc, JsonPrimitive(mdoc.issuerSigned.toMapElement().toCBOR().encodeToBase64Url()),
+//            customParameters = mapOf("credential_encoding" to JsonPrimitive("issuer-signed"))
+//        )
+//    }
+
 }
